% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_columns}
\alias{check_columns}
\title{Internal function intended to throw an error if a needed
column isn't found
TODO: add info about data, function}
\usage{
check_columns(df, cols)
}
\description{
Internal function intended to throw an error if a needed
column isn't found
TODO: add info about data, function
}
