% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_outputs}
\alias{check_outputs}
\title{Internal function to check for output files}
\usage{
check_outputs(runs, prefix, file_count = 10)
}
\arguments{
\item{runs}{Data Table with run information}
}
\value{
Data Table of ReEDS scenarios with boolean 'outputs' column indicating the presence of output csv files
}
\description{
Internal function to check for output files
}
