% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_techs}
\alias{format_techs}
\title{Format techs}
\usage{
format_techs(
  df,
  tech_col = "i",
  mapping = NULL,
  col_from = "raw",
  col_to = "agg1",
  order_techs = T,
  tech_order = NULL
)
}
\arguments{
\item{df}{data frame or table with generator technologies to map in a column specified by \code{tech_col}}

\item{tech_col}{column from \code{df} with tech names (default = \code{i})}

\item{mapping}{data frame or table with mapping of old to new techs; by default uses built in \code{tech_map}}

\item{col_from}{column from \code{mapping} with old tech names (default = \code{raw})}

\item{col_to}{column from \code{mapping} with new tech names (default = \code{agg1})`)}

\item{order_techs}{whether to set tech order using factor levels (default = \code{TRUE})}

\item{tech_order}{named vector of techs in order to set for levels}
}
\value{
data frame or table with 'tech' column added
}
\description{
Recategorizes data by generator techs based on a tech mapping file. Outputs the same data with a new column named \code{tech} that includes the new mapping.
Note that the mapping function will automatically convert the mapping and target technology values to lower case.
}
