% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{inflate_cost_data}
\alias{inflate_cost_data}
\title{Inflate cost data}
\usage{
inflate_cost_data(
  runs,
  cost_df,
  year_from = 2004,
  year_to = NULL,
  cols = NULL,
  drop_original = T
)
}
\arguments{
\item{runs}{datatable of runs to load system costs, generated from run_summary()}

\item{year_from}{integer of original dollar year of data (default: 2004)}

\item{year_to}{integer of dollar year to conver to}

\item{cols}{column names to inflate}

\item{drop_original}{if TRUE drop original dollar year data}

\item{df}{data with cost information}
}
\description{
Inflate cost data
}
