% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{list_bokeh_fields}
\alias{list_bokeh_fields}
\title{List bokeh report fields}
\usage{
list_bokeh_fields(runs, runval = NULL, report = "reeds-report")
}
\arguments{
\item{runs}{data table of runs to search for bokeh report}

\item{runval}{optional parameter to specify name of specific run in runs to look at}

\item{report}{optional parameter to specify bokeh report to look for (default: reeds-report)}
}
\description{
Function to list fields in the report.xlsx file associated with a bokeh report
}
\examples{
list_bokeh_fields(runs)
list_bokeh_fields(runs, runval="runname")
list_bokeh_fields(runs, report="reeds-report-expanded")

}
