% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{list_bokeh_reports}
\alias{list_bokeh_reports}
\title{List bokeh reports}
\usage{
list_bokeh_reports(runs, runval = NULL)
}
\arguments{
\item{runs}{data table of runs to search for bokeh report}

\item{runval}{optional parameter to specify name of specific run in runs to look at}
}
\description{
Function to list available boke reports from a set of runs
}
\examples{
list_bokeh_reports(runs)
list_bokeh_reports(runs, runval="runname")

}
