% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{list_outputs}
\alias{list_outputs}
\title{TODO: complete description here}
\usage{
list_outputs(runs, runval = NULL, folder = "outputs", bokeh = F)
}
\arguments{
\item{runs}{Data Table of runs to list outputs.}
}
\value{
Data Table with results by run name
}
\description{
TODO: complete description here
}
