% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{list_runs}
\alias{list_runs}
\title{List ReEDS runs}
\usage{
list_runs(path, batch = NULL)
}
\arguments{
\item{path}{path to a set of ReEDS runs}

\item{batch}{optional prefix to filter runs; can be specified as a regex pattern}
}
\description{
Prints the names of any ReEDS runs in a given run folder. A batch name
can also be specified to filter to select runs.
}
\examples{
list_runs("path/to/ReEDS-2.0/runs")
list_runs("path/to/ReEDS-2.0/runs", "20241129_testrun")
}
