% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_bokeh_data}
\alias{load_bokeh_data}
\title{Load outputs from bokeh reports}
\usage{
load_bokeh_data(runs, result_name, report_name = "reeds-report")
}
\arguments{
\item{runs}{data table of runs to load outputs}

\item{filename}{filename to load}

\item{folder}{optional parameter to specify folder where filename resides (default: outputs)}

\item{newcolname}{optional parameter to rename value column}
}
\value{
data table with results by run name
}
\description{
Function to collect bokeh report data from a set of ReEDS runs. Defaults to data from "reeds-report"
but be used to extract values from any bokeh report.
}
\examples{
avgcost <- load_bokeh_data(runs, "18_National Average Electricity")

}
