% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_meta}
\alias{load_meta}
\title{Load run metadata}
\usage{
load_meta(runs, summarize = FALSE, skip_lines = 3)
}
\arguments{
\item{runs}{data table object with run and path information, generated from run_summary()}

\item{summarize}{if TRUE summarizes run time and last processing step}

\item{skip_lines}{number of lines to skip in the metadata header}
}
\value{
data table object of run metadata
}
\description{
Load and format meta.csv files from list of runs. Can provide detailed runtime by step or a
summary of total runtime and the last processing step for each run.
}
