% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_reeds_file}
\alias{load_reeds_file}
\title{Load a ReEDS run file}
\usage{
load_reeds_file(path, run_name, header)
}
\arguments{
\item{path}{path to file to load}

\item{run_name}{name of run}
}
\description{
Load a ReEDS run file
}
