% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_run_data}
\alias{load_run_data}
\title{Loads outputs from ReEDS runs; for outputs with generator techs also performs some formatting}
\usage{
load_run_data(runs, filename, folder = "outputs", newcolname = NULL)
}
\arguments{
\item{runs}{Data Table of runs to load outputs.}

\item{output_file}{Filename of output to load.}
}
\value{
Data Table with results by run name
}
\description{
Loads outputs from ReEDS runs; for outputs with generator techs also performs some formatting
}
\examples{
scens <- list_runs("path/to/ReEDS-2.0/runs", "20230416")
scens <- list_runs("path/to/ReEDS-2.0/runs", "20230416|20230422")
}
