% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_run_summary}
\alias{load_run_summary}
\title{Summarize ReEDS runs}
\usage{
load_run_summary(path, batch = "")
}
\arguments{
\item{path}{Path to a set of ReEDS runs}

\item{batch}{optional prefix to filter runs; can be specified as a regex pattern}
}
\value{
data table of ReEDS scenarios and their paths
}
\description{
Function to summarize metadata on set of ReEDS run in a given run folder, with the option to subset to runs that match
a specified batch prefix. The prefix can be specified as a regex string if runs from multiple batches
are of interest.
}
\details{
The function returns a data table with the following information on the runs:
\itemize{
\item run name (dropping any batch prefix if supplied)
\item batch name (if supplied)
\item whether files were detected in the 'outputs' folder
\item the last ReEDS step that completed, as listed in meta.csv
\item the run time, as listed in meta.csv
\item full path to the run
}
}
\examples{
runs <- load_run_summary("path/to/ReEDS-2.0/runs", "20230416")

# runs from multiple batches can be matched by using regex expressions
runs <- load_run_summary("path/to/ReEDS-2.0/runs", "20230416|20230422")

}
