% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_system_cost}
\alias{load_system_cost}
\title{Load system cost}
\usage{
load_system_cost(runs)
}
\arguments{
\item{runs}{datatable of runs to load system costs, generated from run_summary()}
}
\value{
data table with system costs
}
\description{
See 'pre_systemcost' function in ReEDS bokehpivot.
}
