% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{set_tech_map}
\alias{set_tech_map}
\title{Load tech mapping}
\usage{
set_tech_map(path = NULL, reeds_default = T, filename = "tech_map.csv")
}
\arguments{
\item{path}{path to ReEDS repository or a directory with a custom tech mapping}

\item{reeds_default}{Boolean variable indicating whether to use the ReEDS default tech mapping in bokehpivot (defaults to True)}

\item{filename}{optional parameter for specifying filename of custome tech mapping}
}
\value{
dataframe of tech categories
}
\description{
Load mapping for generator techs. Defaults to taking bokehpivot mapping in "tech_map.csv" from a specified ReEDS folder,
but can also be used to load any custom mapping.
}
