% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{set_tech_plot_default}
\alias{set_tech_plot_default}
\title{Set default ggplot settings for tech plots}
\usage{
set_tech_plot_default()
}
\description{
Set ggplot2 theme settings for bar charts
}
\examples{
set_tech_plot_default()

}
