% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{set_tech_style}
\alias{set_tech_style}
\title{Set tech style}
\usage{
set_tech_style(path, filename = "tech_style.csv", custom = F)
}
\arguments{
\item{path}{path to ReEDS repository or a directory with a custom tech style}

\item{filename}{typically "tech_style.csv" but can optionally be specified as any custom file (default = "tech_style.csv")}

\item{custom}{flag indicating whether to use a custom tech style or the ReEDS default in bokehpivot (default = FALSE)}
}
\value{
named list of techs and colors
}
\description{
Set the color style scheme file for generator techs. Should correspond to the techs specified in the tech mapping file (see set_tech_map)
Defaults to taking bokehpivot mapping in "tech_style.csv" when passed a path to ReEDS,
but can also be used to load any custom style file
}
\examples{
tech_style <- set_tech_style("path/to/ReEDS-2.0")
tech_map <- set_tech_map("path/to/customstyle", custom=T, file="custom_tech_style.csv")

}
