% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tech_colors}
\alias{tech_colors}
\title{ReEDS technology colors}
\format{
\subsection{\code{tech_colors}}{

Named character vector of technology colors.
}
}
\usage{
tech_colors
}
\description{
Named character vector with standard technology colors. The order of the entries can
can be used to set levels for a typical dispatch stack plot.
}
\keyword{datasets}
