% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tech_map}
\alias{tech_map}
\title{ReEDS technology map}
\format{
\subsection{\code{tech_map}}{

A data frame with 44 rows and 3 columns:
\describe{
\item{raw}{Technology name in ReEDS (not including class and mod/max designations)}
\item{agg1, agg2}{Two alternative options for technology categories}
}
}
}
\usage{
tech_map
}
\description{
Maps ReEDS i set technologies to aggregated categories.
}
\keyword{datasets}
