% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_techs}
\alias{format_techs}
\title{Format techs}
\usage{
format_techs(df, mapping, col_from = "raw", col_to = "display")
}
\arguments{
\item{path}{path to ReEDS repository or a directory with a custom tech mapping}

\item{reeds_default}{Boolean variable indicating whether to use the ReEDS default tech mapping in bokehpivot (defaults to True)}

\item{filename}{optional parameter for specifying filename of custome tech mapping}
}
\value{
Data Table of tech categories
}
\description{
Recategorizes generator techs based on mapping.
}
