% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{list_outputs}
\alias{list_outputs}
\title{List output files}
\usage{
list_outputs(runs, folder = "outputs", runval = NULL, keyword = NULL)
}
\arguments{
\item{runs}{datatable of runs to list outputs, generated from run_summary()}

\item{folder}{optional parameter to specify folder to look for files (default: outputs)}

\item{runval}{optional parameter to specify name of specific run in runs to look at}
}
\description{
Prints a list of files in a folder of runs. Defaults to looking at the outputs folder of the first run specified.
}
\examples{
list_outputs(runs)
list_outputs(runs, runval="runname")
list_outputs(runs, folder="inputs_case")

}
