% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{list_runs}
\alias{list_runs}
\title{List ReEDS runs}
\usage{
list_runs(path)
}
\arguments{
\item{path}{Path to a set of ReEDS runs}
}
\description{
Function to discover a set of ReEDS run in a given run folder.
}
\examples{
runs <- list_runs("path/to/ReEDS-2.0/runs")

}
