% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_meta}
\alias{load_meta}
\title{Internal function to get runtime from meta.csv file}
\usage{
load_meta(runs)
}
\arguments{
\item{runs}{Data Table with run information}
}
\value{
Data Table of ReEDS scenarios with the total runtime and last time step
}
\description{
Internal function to get runtime from meta.csv file
}
